requirements.txt
pandas>=1.1.5
numpy>=1.20.1
plotly==4.14.3
streamlit>=0.84.1
google-cloud-firestore
google-auth
hickle
python-box